Refactor stylesheet handling: consolidate stylesheet parsing and apply styles to UINodes
