+++ src/graphics/ui/elements/Container.cpp
@@ -4,7 +4,6 @@
 #include <iostream>
 #include <utility>
 
-#include "../style/StylesheetParser.h"
 #include "graphics/core/Batch2D.hpp"
 #include "graphics/core/DrawContext.hpp"
 
@@ -119,189 +118,19 @@ void Container::setScrollable(bool flag) {
     scrollable = flag;
 }
 
-static std::string trim(const std::string& s) {
-    if (s.empty()) return s;
-
-    size_t start = s.find_first_not_of(" \t\n\r");
-    size_t end = s.find_last_not_of(" \t\n\r");
-
-    if (start == std::string::npos) return "";
-    return s.substr(start, end - start + 1);
-}
-
-std::unordered_map<std::string, style::value> calculateStylesheet(
-    const UINode& node, 
-    const std::vector<StylesheetRule>& rules
-) {
-    struct AppliedRule {
-        const StylesheetRule* rule;
-        int specificity;
-        size_t index;
-    };
-    
-    std::vector<AppliedRule> appliedRules;
-    appliedRules.reserve(rules.size());
-    
-    for (size_t i = 0; i < rules.size(); ++i) {
-        const auto& rule = rules[i];
-        
-        // Разбиваем составные селекторы (button, .menu)
-        std::vector<std::string> selectors;
-        size_t start = 0;
-        size_t comma;
-        do {
-            comma = rule.selector.find(',', start);
-            size_t end = (comma == std::string::npos) ? rule.selector.length() : comma;
-            std::string sel = trim(rule.selector.substr(start, end - start));
-            if (!sel.empty()) {
-                selectors.push_back(sel);
-            }
-            start = comma + 1;
-        } while (comma != std::string::npos);
-        
-        bool selectorMatches = false;
-        
-        for (const auto& selector : selectors) {
-            // Проверяем, является ли селектор универсальным *
-            if (selector == "*") {
-                // Универсальный селектор применяется ко всем элементам
-                selectorMatches = true;
-                
-                // Специфичность для * = 0
-                appliedRules.push_back({&rule, 0, i});
-                break; // Другие селекторы в этом правиле не проверяем
-            }
-            
-            // Парсим обычный селектор
-            std::string tag, id, classname;
-            std::vector<std::string> pseudoClasses;
-            
-            size_t pos = 0;
-            while (pos < selector.length()) {
-                if (std::isspace(selector[pos])) {
-                    pos++;
-                    continue;
-                }
-                
-                if (selector[pos] == '#') {
-                    pos++;
-                    size_t end = selector.find_first_of(".:#[] ", pos);
-                    if (end == std::string::npos) end = selector.length();
-                    id = selector.substr(pos, end - pos);
-                    pos = end;
-                } 
-                else if (selector[pos] == '.') {
-                    pos++;
-                    size_t end = selector.find_first_of(".:#[] ", pos);
-                    if (end == std::string::npos) end = selector.length();
-                    classname = selector.substr(pos, end - pos);
-                    pos = end;
-                } 
-                else if (selector[pos] == ':') {
-                    pos++;
-                    size_t end = selector.find_first_of(".:#[] ", pos);
-                    if (end == std::string::npos) end = selector.length();
-                    pseudoClasses.push_back(selector.substr(pos, end - pos));
-                    pos = end;
-                } 
-                else {
-                    size_t end = selector.find_first_of(".:#[] ", pos);
-                    if (end == std::string::npos) end = selector.length();
-                    tag = selector.substr(pos, end - pos);
-                    pos = end;
-                }
-            }
-            
-            // Проверяем соответствие узлу
-            bool matches = true;
-            
-            // Проверяем тег (если указан и не *)
-            if (!tag.empty() && tag != "*" && node.getName() != tag) {
-                matches = false;
-            }
-            
-            // Проверяем ID (если указан)
-            if (!id.empty() && node.getId() != id) {
-                matches = false;
-            }
-            
-            // Проверяем класс (если указан)
-            if (!classname.empty() && node.getClassname() != classname) {
-                matches = false;
-            }
-            
-            // Проверяем все псевдоклассы
-            for (const auto& pseudo : pseudoClasses) {
-                if (pseudo == "hover" && !node.isHover()) {
-                    matches = false;
-                    break;
-                } 
-                else if (pseudo == "active" && !node.isPressed()) {
-                    matches = false;
-                    break;
-                } 
-                else if (pseudo == "focus" && !node.isFocused()) {
-                    matches = false;
-                    break;
-                } 
-                else if (pseudo == "disabled" && node.isEnabled()) {
-                    matches = false;
-                    break;
-                }
-            }
-            
-            if (matches) {
-                selectorMatches = true;
-                
-                // Вычисляем специфичность селектора
-                int specificity = 0;
-                
-                if (!id.empty()) specificity += 100;
-                if (!classname.empty()) specificity += 10;
-                if (!tag.empty() && tag != "*") specificity += 1;
-                specificity += pseudoClasses.size() * 10;
-                
-                appliedRules.push_back({&rule, specificity, i});
-                break; // Если один из составных селекторов подходит, остальные не проверяем
-            }
+std::string stylesheet = R"(
+        label {
+            color: #ff0000;
         }
-    }
-    
-    // Сортируем правила по специфичности
-    std::sort(appliedRules.begin(), appliedRules.end(), 
-        [](const AppliedRule& a, const AppliedRule& b) {
-            if (a.specificity != b.specificity) {
-                return a.specificity > b.specificity;
-            }
-            return a.index < b.index; // Ранние правила имеют приоритет при одинаковой специфичности
-        });
-    
-    // Применяем стили
-    std::unordered_map<std::string, style::value> styles;
-    
-    for (const auto& applied : appliedRules) {
-        for (const auto& [property, value] : applied.rule->declarations) {
-            // Более специфичные правила переопределяют менее специфичные
-            if (styles.find(property) == styles.end()) {
-                styles[property] = value;
-            }
+        button:hover {
+            color: #22222288;
         }
-    }
-    
-    return styles;
-}
+        button {
+            color: #11111188;
+        }
+    )";
 
-std::string stylesheet_src = R"(
-    button:hover {
-        color: #575757ff;
-    }
-    button { 
-        color: #000000;
-    }
-    * {
-        color: #00ff00;
-    }
-)";
+std::vector<StylesheetRule> rules = StylesheetParser::parse(stylesheet);
 
 void Container::draw(const DrawContext& pctx, const Assets& assets) {
     glm::vec2 pos = calcPos();
@@ -318,25 +147,7 @@ void Container::draw(const DrawContext& pctx, const Assets& assets) {
         );
         for (const auto& node : nodes) {
             if (node->isVisible()) {
-                if (node->getName() != "") {
-                    std::vector<StylesheetRule> styles = StylesheetParser::parse(stylesheet_src);
-                    std::unordered_map<std::string, style::value> node_styles = calculateStylesheet(*node, styles);
-                    for (const auto& [property, value] : node_styles) {
-                        switch (getStylePropertyType(property))
-                        {
-                        case StyleProperty::COLOR:
-                            node->setColor( value.asColor() );
-                            break;
-                        case StyleProperty::MARGIN:
-                            node->setMargin( value.asVec4() );
-                            break;
-
-                        default:
-                            break;
-                        }
-                    }
-                }
-
+                node->applyStylesheet(rules);
                 node->draw(pctx, assets);
             }
         }
