+++ src/graphics/ui/elements/UINode.cpp
@@ -437,3 +437,176 @@ void UINode::setName(std::string _name) {
 std::string UINode::getName() const {
     return name;
 }
+
+void UINode::applyStylesheet(const std::vector<StylesheetRule>& rules) {
+    std::unordered_map<std::string, style::value> styles = calculateStylesheet(rules);
+    for (const auto& [property, value] : styles) {
+        switch (getStylePropertyType(property))
+        {
+        case StyleProperty::COLOR:
+            setColor( value.asColor() );
+            break;
+        case StyleProperty::MARGIN:
+            setMargin( value.asVec4() );
+            break;
+        case StyleProperty::Z_INDEX:
+            setZIndex( value.asInt() );
+            break;
+        default:
+            break;
+        }
+    }
+}
+
+std::unordered_map<std::string, style::value> UINode::calculateStylesheet(const std::vector<StylesheetRule>& rules) {
+    std::vector<AppliedRule> appliedRules;
+    appliedRules.reserve(rules.size());
+    
+    for (size_t i = 0; i < rules.size(); ++i) {
+        const auto& rule = rules[i];
+        
+        // Разбиваем составные селекторы (button, .menu)
+        std::vector<std::string> selectors;
+        size_t start = 0;
+        size_t comma;
+        do {
+            comma = rule.selector.find(',', start);
+            size_t end = (comma == std::string::npos) ? rule.selector.length() : comma;
+            std::string sel = trim(rule.selector.substr(start, end - start));
+            if (!sel.empty()) {
+                selectors.push_back(sel);
+            }
+            start = comma + 1;
+        } while (comma != std::string::npos);
+        
+        bool selectorMatches = false;
+        
+        for (const auto& selector : selectors) {
+            // Проверяем, является ли селектор универсальным *
+            if (selector == "*") {
+                // Универсальный селектор применяется ко всем элементам
+                selectorMatches = true;
+                
+                // Специфичность для * = 0
+                appliedRules.push_back({&rule, 0, i});
+                break; // Другие селекторы в этом правиле не проверяем
+            }
+            
+            // Парсим обычный селектор
+            std::string tag, id, classname;
+            std::vector<std::string> pseudoClasses;
+            
+            size_t pos = 0;
+            while (pos < selector.length()) {
+                if (std::isspace(selector[pos])) {
+                    pos++;
+                    continue;
+                }
+                
+                if (selector[pos] == '#') {
+                    pos++;
+                    size_t end = selector.find_first_of(".:#[] ", pos);
+                    if (end == std::string::npos) end = selector.length();
+                    id = selector.substr(pos, end - pos);
+                    pos = end;
+                } 
+                else if (selector[pos] == '.') {
+                    pos++;
+                    size_t end = selector.find_first_of(".:#[] ", pos);
+                    if (end == std::string::npos) end = selector.length();
+                    classname = selector.substr(pos, end - pos);
+                    pos = end;
+                } 
+                else if (selector[pos] == ':') {
+                    pos++;
+                    size_t end = selector.find_first_of(".:#[] ", pos);
+                    if (end == std::string::npos) end = selector.length();
+                    pseudoClasses.push_back(selector.substr(pos, end - pos));
+                    pos = end;
+                } 
+                else {
+                    size_t end = selector.find_first_of(".:#[] ", pos);
+                    if (end == std::string::npos) end = selector.length();
+                    tag = selector.substr(pos, end - pos);
+                    pos = end;
+                }
+            }
+            
+            // Проверяем соответствие узлу
+            bool matches = true;
+            
+            // Проверяем тег (если указан и не *)
+            if (!tag.empty() && tag != "*" && this->getName() != tag) {
+                matches = false;
+            }
+            
+            // Проверяем ID (если указан)
+            if (!id.empty() && this->getId() != id) {
+                matches = false;
+            }
+            
+            // Проверяем класс (если указан)
+            if (!classname.empty() && this->getClassname() != classname) {
+                matches = false;
+            }
+            
+            // Проверяем все псевдоклассы
+            for (const auto& pseudo : pseudoClasses) {
+                if (pseudo == "hover" && !this->isHover()) {
+                    matches = false;
+                    break;
+                } 
+                else if (pseudo == "active" && !this->isPressed()) {
+                    matches = false;
+                    break;
+                } 
+                else if (pseudo == "focus" && !this->isFocused()) {
+                    matches = false;
+                    break;
+                } 
+                else if (pseudo == "disabled" && this->isEnabled()) {
+                    matches = false;
+                    break;
+                }
+            }
+            
+            if (matches) {
+                selectorMatches = true;
+                
+                // Вычисляем специфичность селектора
+                int specificity = 0;
+                
+                if (!id.empty()) specificity += 100;
+                if (!classname.empty()) specificity += 10;
+                if (!tag.empty() && tag != "*") specificity += 1;
+                specificity += pseudoClasses.size() * 10;
+                
+                appliedRules.push_back({&rule, specificity, i});
+                break; // Если один из составных селекторов подходит, остальные не проверяем
+            }
+        }
+    }
+    
+    // Сортируем правила по специфичности
+    std::sort(appliedRules.begin(), appliedRules.end(), 
+        [](const AppliedRule& a, const AppliedRule& b) {
+            if (a.specificity != b.specificity) {
+                return a.specificity > b.specificity;
+            }
+            return a.index < b.index; // Ранние правила имеют приоритет при одинаковой специфичности
+        });
+    
+    // Применяем стили
+    std::unordered_map<std::string, style::value> styles;
+    
+    for (const auto& applied : appliedRules) {
+        for (const auto& [property, value] : applied.rule->declarations) {
+            // Более специфичные правила переопределяют менее специфичные
+            if (styles.find(property) == styles.end()) {
+                styles[property] = value;
+            }
+        }
+    }
+    
+    return styles;
+}
